WP-77 Move auth folder to services
